+++ core-ui/src/components/Lambdas/LambdasList/LambdasList.js
@@ -1,46 +1,18 @@
 import React from 'react';
 import LuigiClient from '@luigi-project/client';
 
-import { GenericList, Labels, StatusBadge } from 'react-shared';
+import { GenericList, Labels } from 'react-shared';
 import { Link } from 'fundamental-react';
+import { LambdaStatusBadge } from '../LambdaStatusBadge/LambdaStatusBadge';
 
 import { useDeleteLambda } from 'components/Lambdas/gql/hooks/mutations';
 
-import { statusType } from 'components/Lambdas/helpers/lambdas';
-import { formatMessage } from 'components/Lambdas/helpers/misc';
 import { prettyRuntime } from 'components/Lambdas/helpers/runtime';
 import {
   LAMBDAS_LIST,
-  LAMBDA_PHASES,
-  LAMBDA_ERROR_PHASES,
   ERRORS,
   REFETCH_LAMBDAS_TIMEOUT,
 } from 'components/Lambdas/constants';
-
-function LambdaStatusBadge({ status }) {
-  const statusPhase = status.phase;
-
-  const texts = LAMBDA_PHASES[statusPhase];
-  let badgeType = statusType(statusPhase);
-  if (badgeType === 'info') {
-    badgeType = undefined;
-  }
-
-  let tooltipText;
-
-  if (LAMBDA_ERROR_PHASES.includes(statusPhase)) {
-    const formattedError = formatMessage(LAMBDA_PHASES.ERROR_SUFFIX, {
-      error: status.message,
-    });
-    tooltipText = `${texts.MESSAGE} ${formattedError}`;
-  }
-
-  return (
-    <StatusBadge tooltipContent={tooltipText} type={badgeType}>
-      {texts.TITLE}
-    </StatusBadge>
-  );
-}
 
 const headerRenderer = () => ['Name', 'Runtime', 'Labels', 'Status'];
 const textSearchProperties = ['name', 'prettyRuntime', 'status.phase'];
